moved all of the manual api requests into the api layer, fixed some potential overflow issues, moved auth into the api layer
the general structure is making a bit more sense now! closes #60

+++ types/gamified_lesson_plan.go
@@ -1,11 +1,11 @@
 package types
 
 type GamifiedLessonPlan struct {
-	Id           int
+	Id           uint64
 	Name         string
 	Desc         string
 	Author       string
 	Category     string
-	GamePlotId   int
+	GamePlotId   uint64
 	ExternConfig string
 }
